Refactor order namespace and add javadoc comments
