+++ src/main/kotlin/com/hhandoko/luwak/api/order/OrderData.kt
@@ -1,5 +1,5 @@
 /**
- * File     : OrdersResponse.kt
+ * File     : OrderData.kt
  * License  :
  *   Copyright (c) 2017 Herdy Handoko
  *
@@ -15,6 +15,12 @@
  *   See the License for the specific language governing permissions and
  *   limitations under the License.
  */
-package com.hhandoko.luwak.api
+package com.hhandoko.luwak.api.order
 
-data class OrdersResponse(val results: List<OrderData>, val success: Boolean = true)
+/**
+ * Order response data.
+ *
+ * @param ref The unique order reference.
+ * @param name The order.
+ */
+data class OrderData(val ref: String, val name: String)
