+++ src/main/kotlin/com/hhandoko/luwak/resource/OrderResource.kt
@@ -17,31 +17,42 @@
  */
 package com.hhandoko.luwak.resource
 
+import java.util.*
 import javax.ws.rs.GET
 import javax.ws.rs.Path
 import javax.ws.rs.PathParam
 import javax.ws.rs.Produces
 import javax.ws.rs.core.MediaType
 
-import com.hhandoko.luwak.api.OrderData
-import com.hhandoko.luwak.api.OrderResponse
-import com.hhandoko.luwak.api.OrdersResponse
-import java.util.*
+import com.hhandoko.luwak.api.order.OrderData
+import com.hhandoko.luwak.api.order.OrderResponse
+import com.hhandoko.luwak.api.order.OrdersResponse
 
+/**
+ * Coffee orders endpoint.
+ */
 @Path("/orders")
 @Produces(MediaType.APPLICATION_JSON)
 class OrderResource {
 
+    /**
+     * Get all orders.
+     */
     @GET
     fun get(): OrdersResponse {
         val order = OrderData(UUID.randomUUID().toString(), "Coffee")
         return OrdersResponse(listOf(order))
     }
 
+    /**
+     * Get a single order by its reference.
+     *
+     * @param ref The unique order reference.
+     */
     @GET
     @Path("/{ref}")
-    fun getOne(@PathParam("ref") id: String): OrderResponse {
-        val order = OrderData(id, "Coffee")
+    fun getOne(@PathParam("ref") ref: String): OrderResponse {
+        val order = OrderData(ref, "Coffee")
         return OrderResponse(order)
     }
